Add skip button and check for torch on to heart rate measurement
For some reason the torch didn’t turn on at the end of the 12 minute run. Moving turning on the torch back to before all the other settings and also checking if the lens isn’t covered *AND* add a skip button just in case so that a user doesn’t lose all their data.  :(

S